added typesafe envs & reusable function
