+++ app/api/leaderboard/functions.ts
@@ -1,6 +1,10 @@
 import { LeaderboardSortKey } from "@/app/leaderboard/Leaderboard";
 import { getContributors } from "@/lib/api";
 import { Highlights } from "@/lib/types";
+import { ReleasesResponse } from "@/lib/types";
+import { Repository } from "@/lib/types";
+import { Release } from "@/lib/types";
+import { env } from "@/env.mjs";
 
 export type LeaderboardAPIResponse = {
   user: {
@@ -81,3 +85,65 @@ export const getLeaderboardData = async (
     };
   });
 };
+
+export default async function fetchGitHubReleases(
+  sliceLimit: number,
+): Promise<Release[]> {
+  const response = await fetch("https://api.github.com/graphql", {
+    method: "POST",
+    headers: {
+      Authorization: `Bearer ${env.GITHUB_PAT}`,
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({
+      query: `{
+        organization(login: "${env.NEXT_PUBLIC_GITHUB_ORG}") {
+          repositories(first: 100) {
+            nodes {
+              name
+              releases(first: 10, orderBy: {field: CREATED_AT, direction: DESC}) {
+                nodes {
+                  name
+                  createdAt
+                  description
+                  url
+                  author {
+                    login
+                    avatarUrl
+                  }
+                  mentions (first: 10) {
+                    nodes {
+                      login 
+                      avatarUrl
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      }`,
+    }),
+  });
+
+  if (!response.ok) {
+    throw new Error(`HTTP error! status: ${response.status}`);
+  }
+
+  const json = (await response.json()) as ReleasesResponse;
+  const repositories: Repository[] = json.data.organization.repositories.nodes;
+  const allReleases: Release[] = [];
+
+  for (const repository of repositories) {
+    for (const release of repository.releases.nodes) {
+      release.repository = repository.name;
+      allReleases.push(release);
+    }
+  }
+
+  const sortedReleases = allReleases.sort(
+    (a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime(),
+  );
+
+  return sortedReleases.slice(0, sliceLimit);
+}
